user authorization completed, update production env

+++ app/controllers/bookmarks_controller.rb
@@ -1,5 +1,6 @@
 class BookmarksController < ApplicationController
   before_action :authenticate_user!
+  rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
 
   def show
     @bookmark = Bookmark.find(params[:id])
@@ -13,10 +14,13 @@ class BookmarksController < ApplicationController
   def create
     @topic = Topic.find(params[:topic_id])
     @bookmark = @topic.bookmarks.create(bookmark_params)
+    @bookmark.user = current_user
 
     if @bookmark.save
+      flash[:notice] = "Bookmark was saved."
       redirect_to @bookmark.topic
     else
+      flash.now[:alert] = "There was an error saving the bookmark. Please try again."
       redirect_to @bookmark.topic
     end
   end
@@ -29,22 +33,28 @@ class BookmarksController < ApplicationController
     @topic = Topic.find(params[:topic_id])
     @bookmark = Bookmark.find(params[:id])
     @bookmark.assign_attributes(bookmark_params)
+    authorize @bookmark
 
     if @bookmark.save
+      flash[:notice] = "Bookmark was updated."
       redirect_to topic_path(@topic)
     else
-      redirect_to @bookmark.topic
+      flash.now[:alert] = "There was an error updating the bookmark. Please try again."
+      redirect_to topic_path(@topic)
     end
   end
 
   def destroy
     @topic = Topic.find(params[:topic_id])
     @bookmark = Bookmark.find(params[:id])
+    authorize @bookmark
 
     if @bookmark.destroy
+      flash[:notice] = "Bookmark was deleted."
       redirect_to topic_path(@topic)
     else
-      redirect_to @bookmark.topic
+      flash.now[:alert] = "There was an error deleting the bookmark. Please try again."
+      redirect_to topic_path(@topic)
     end
   end
 
@@ -52,5 +62,10 @@ class BookmarksController < ApplicationController
 
   def bookmark_params
     params.require(:bookmark).permit(:url)
+  end
+
+  def user_not_authorized
+    flash[:alert] = "You are not authorized to perform this action."
+    redirect_to(request.referrer || root_path)
   end
 end
